Moving the workaround for nose's stdout-capturing to utils.py 	Make use of it in test_annexrepo.py
